+++ cmd/seer-q/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"os"
 
-	"github.com/chris-xu0321/Midaz-cli/internal/cli"
+	"github.com/SparkssL/Midaz-cli/internal/cli"
 )
 
 func main() {