+++ internal/cmd/claims/claims.go
@@ -3,7 +3,7 @@ package claims
 import (
 	"net/url"
 
-	"github.com/chris-xu0321/Midaz-cli/internal/cmdutil"
+	"github.com/SparkssL/Midaz-cli/internal/cmdutil"
 	"github.com/spf13/cobra"
 )