+++ internal/cmd/config/config.go
@@ -1,7 +1,7 @@
 package config
 
 import (
-	"github.com/chris-xu0321/Midaz-cli/internal/cmdutil"
+	"github.com/SparkssL/Midaz-cli/internal/cmdutil"
 	"github.com/spf13/cobra"
 )