+++ internal/cmd/config/set.go
@@ -3,9 +3,9 @@ package config
 import (
 	"slices"
 
-	"github.com/chris-xu0321/Midaz-cli/internal/cmdutil"
-	cfgpkg "github.com/chris-xu0321/Midaz-cli/internal/config"
-	"github.com/chris-xu0321/Midaz-cli/internal/output"
+	"github.com/SparkssL/Midaz-cli/internal/cmdutil"
+	cfgpkg "github.com/SparkssL/Midaz-cli/internal/config"
+	"github.com/SparkssL/Midaz-cli/internal/output"
 	"github.com/spf13/cobra"
 )