+++ internal/cmd/doctor/doctor.go
@@ -5,10 +5,10 @@ import (
 	"os"
 	"time"
 
-	"github.com/chris-xu0321/Midaz-cli/internal/client"
-	"github.com/chris-xu0321/Midaz-cli/internal/cmdutil"
-	"github.com/chris-xu0321/Midaz-cli/internal/config"
-	"github.com/chris-xu0321/Midaz-cli/internal/output"
+	"github.com/SparkssL/Midaz-cli/internal/client"
+	"github.com/SparkssL/Midaz-cli/internal/cmdutil"
+	"github.com/SparkssL/Midaz-cli/internal/config"
+	"github.com/SparkssL/Midaz-cli/internal/output"
 	"github.com/spf13/cobra"
 )