+++ internal/cmd/schema/schema.go
@@ -1,8 +1,8 @@
 package schema
 
 import (
-	"github.com/chris-xu0321/Midaz-cli/internal/cmdutil"
-	"github.com/chris-xu0321/Midaz-cli/internal/output"
+	"github.com/SparkssL/Midaz-cli/internal/cmdutil"
+	"github.com/SparkssL/Midaz-cli/internal/output"
 	"github.com/spf13/cobra"
 )