+++ internal/cmd/topics/topics.go
@@ -1,7 +1,7 @@
 package topics
 
 import (
-	"github.com/chris-xu0321/Midaz-cli/internal/cmdutil"
+	"github.com/SparkssL/Midaz-cli/internal/cmdutil"
 	"github.com/spf13/cobra"
 )